fixed some typos and issues
